Move out of cmd folder (#4450)
