+++ examples/nullchain/connection.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	config "code.vegaprotocol.io/vega/cmd/examples/nullchain/config"
+	config "code.vegaprotocol.io/vega/examples/nullchain/config"
 	"github.com/pkg/errors"
 
 	datanode "code.vegaprotocol.io/protos/data-node/api/v1"
