+++ examples/nullchain/timeforward.go
@@ -9,7 +9,7 @@ import (
 	"net/http"
 	"time"
 
-	config "code.vegaprotocol.io/vega/cmd/examples/nullchain/config"
+	config "code.vegaprotocol.io/vega/examples/nullchain/config"
 )
 
 var ErrTimeForward = errors.New("time forward failed")
