+++ examples/nullchain/wallet.go
